feat!: Improved iterators on OutgoingPacket
Closes #27
